+++ tdp/cli/commands/status/edit.py
@@ -68,7 +68,7 @@ def edit(
         )
 
     cluster_variables = ClusterVariables.get_cluster_variables(
-        collections=collections, tdp_vars=vars, validate=validate
+        collections=collections, tdp_vars_path=vars, validate=validate
     )
     check_services_cleanliness(cluster_variables)
