+++ tdp/cli/commands/status/generate_stales.py
@@ -39,7 +39,7 @@ def generate_stales(
     reconfigured and/or restarted.
     """
     cluster_variables = ClusterVariables.get_cluster_variables(
-        collections=collections, tdp_vars=vars, validate=validate
+        collections=collections, tdp_vars_path=vars, validate=validate
     )
     check_services_cleanliness(cluster_variables)
