+++ tdp/core/conftest.py
@@ -77,7 +77,7 @@ def mock_cluster_variables(
     tmp_path_factory: pytest.TempPathFactory, mock_collections: Collections
 ) -> ClusterVariables:
     return ClusterVariables.initialize_tdp_vars(
-        collections=mock_collections, tdp_vars=tmp_path_factory.mktemp("tdp_vars")
+        collections=mock_collections, tdp_vars_path=tmp_path_factory.mktemp("tdp_vars")
     )
